refactor: moved image description model to data package
