+++ app/src/main/java/org/fossasia/phimpme/accounts/AccountActivity.kt
@@ -17,8 +17,8 @@ import io.realm.RealmResults
 import org.fossasia.phimpme.R
 import org.fossasia.phimpme.base.PhimpmeProgressBarHandler
 import org.fossasia.phimpme.base.ThemedActivity
-import org.fossasia.phimpme.data.AccountDatabase
-import org.fossasia.phimpme.data.DatabaseHelper
+import org.fossasia.phimpme.data.local.AccountDatabase
+import org.fossasia.phimpme.data.local.DatabaseHelper
 import org.fossasia.phimpme.leafpic.util.ThemeHelper
 import org.fossasia.phimpme.utilities.ActivitySwitchHelper
 import org.fossasia.phimpme.base.RecyclerItemClickListner
