--- app/src/main/java/org/fossasia/phimpme/data/DatabaseHelper.java
@@ -1,33 +1,0 @@
-package org.fossasia.phimpme.data;
-
-import io.realm.Realm;
-import io.realm.RealmResults;
-
-/**
- * Created by pa1pal on 10/6/17.
- */
-
-public class DatabaseHelper {
-
-    private Realm realm;
-
-    public DatabaseHelper(Realm realm) {
-        this.realm = realm;
-    }
-
-    public RealmResults<AccountDatabase> fetchAccountDetails(){
-        return realm.where(AccountDatabase.class).findAll();
-    }
-
-    public void deleteSignedOutAccount(String accountName){
-        final RealmResults<AccountDatabase> deletionQueryResult =  realm.where(AccountDatabase.class)
-                .equalTo("name", accountName).findAll();
-
-        realm.executeTransaction(new Realm.Transaction() {
-            @Override
-            public void execute(Realm realm) {
-                deletionQueryResult.deleteAllFromRealm();
-            }
-        });
-    }
-}
